+++ src/entities/notehistory.h
@@ -2,6 +2,7 @@
 #define NOTEHISTORY_H
 
 #include <QTextEdit>
+#include <QDataStream>
 #include "note.h"
 
 
@@ -9,9 +10,12 @@ class NoteHistoryItem
 {
 public:
     explicit NoteHistoryItem(Note *note = NULL, QTextEdit *textEdit = NULL);
+    explicit NoteHistoryItem(QString noteName, int cursorPosition,
+                             float relativeScrollBarPosition);
     friend QDebug operator<<(QDebug dbg, const NoteHistoryItem &item);
     QString getNoteName() const;
     int getCursorPosition() const;
+    float getRelativeScrollBarPosition() const;
     Note getNote();
     bool isNoteValid();
     bool operator ==(const NoteHistoryItem &item) const;
@@ -23,6 +27,11 @@ private:
     float _relativeScrollBarPosition;
     static float getTextEditScrollBarRelativePosition(QTextEdit *textEdit);
 };
+
+// we want to store the class to the settings
+QDataStream &operator<<(QDataStream &out, const NoteHistoryItem &item);
+QDataStream &operator>>(QDataStream &in, NoteHistoryItem &item);
+Q_DECLARE_METATYPE(NoteHistoryItem)
 
 class NoteHistory
 {