+++ src/main.cpp
@@ -157,6 +157,12 @@ bool mainStartupMisc() {
 }
 
 int main(int argc, char *argv[]) {
+    // register NoteHistoryItem so we can store it to the settings
+    // we need to do that before we are accessing QSettings or the
+    // NoteHistoryItem instances in the settings will get destroyed
+    qRegisterMetaTypeStreamOperators<NoteHistoryItem>("NoteHistoryItem");
+    qRegisterMetaType<NoteHistoryItem>("NoteHistoryItem");
+
     QString release = RELEASE;
     bool portable = false;