+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # QOwnNotes Changelog
 
+## 17.03.3
+- after starting QOwnNotes now the last note along with your cursor and 
+  scrollbar position will be restored
+  (for [Issue #537](https://github.com/pbek/QOwnNotes/issues/537))
+    - if that fails the first note in the note list will be opened
+
 ## 17.03.2
 - fixed a scrollbar position warning
 - when pressing <kbd>Tab</kbd> in the note list to switch to the note text